chg: refactored UI module locations
