+++ .travis.yml
@@ -12,9 +12,9 @@ env:
         - DJANGO_VERSION=2.1.4
         - DJANGO_VERSION=2.1.0
         - DJANGO_VERSION=2.1.5
-        - DJANGO_Version=2.1.7
+        - DJANGO_VERSION=2.1.7
 install:
-        - pip install Django==$DJANGO_VERSION markdown MarkdownSuperscript MarkdownSubscript
+        - pip install Django==$DJANGO_VERSION markdown==3.0.1 MarkdownSuperscript MarkdownSubscript
         - pip install pyembed-markdown django_extensions pillow
         - pip install markdown-checklist reportlab qrcode django-email-extras
         - pip install django-helpdesk django-bootstrap4 django-nose coverage
